refactor: features 폴더를 domains로 이관 및 미사용 코드 제거
- features/product 데이터와 상수를 domains/product로 이관
- 상품 관련 constants와 data를 domains/product 내부로 정리
- 기존 constants/discount.js와 features 상수들을 통합하여 중복 제거
- 미사용 features/manual, features/cart 폴더 완전 제거
- 모든 import 경로를 새로운 domains 구조로 업데이트
- 모든 87개 테스트 통과하여 기능 동일성 보장
