+++ src/basic/domains/product/constants.js
@@ -1,0 +1,15 @@
+export const PRODUCT_KEY = Object.freeze({
+	KEYBOARD: "KEYBOARD",
+	MOUSE: "MOUSE",
+	MONITOR_ARM: "MONITOR_ARM",
+	LAPTOP_POUCH: "LAPTOP_POUCH",
+	SPEAKER: "SPEAKER"
+});
+
+export const PRODUCT_IDS = Object.freeze({
+	[PRODUCT_KEY.KEYBOARD]: "p1",
+	[PRODUCT_KEY.MOUSE]: "p2",
+	[PRODUCT_KEY.MONITOR_ARM]: "p3",
+	[PRODUCT_KEY.LAPTOP_POUCH]: "p4",
+	[PRODUCT_KEY.SPEAKER]: "p5"
+});
