--- src/basic/features/manual/components/index.js
@@ -1,3 +1,0 @@
-export * from "./ManualColumn";
-export * from "./ManualOverlay";
-export * from "./ManualToggle";
