--- src/basic/features/product/data/index.js
@@ -1,1 +1,0 @@
-export * from "./products";
